+++ src/main/java/me/tabernerojerry/repository/ILoanRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Repository
 public interface ILoanRepository extends CrudRepository<Loans, Long> {
 
-    List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+    List<Loans> findByCustomerIdOrderByStartDtDesc(Long customerId);
 
 }
